requirements.txt
python-telegram-bot==13.14
python-dotenv==0.21.0
requests==2.28.1
Dialogflow==2.0.0
google-cloud-storage==2.7.0
google-cloud-api-keys==0.5.1